Moved voting connector to its own package
