+++ packages/connector-thegraph-voting/package.json
@@ -1,0 +1,32 @@
+{
+  "name": "plumbery-connector-thegraph-voting",
+  "version": "1.0.0",
+  "license": "AGPL-3.0",
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
+  "scripts": {
+    "build": "yarn clean && yarn compile",
+    "dev": "yarn build --watch",
+    "clean": "rm -rf ./dist && rm -f tsconfig.tsbuildinfo",
+    "compile": "tsc -p tsconfig.json",
+    "lint": "eslint --ext .ts ./src"
+  },
+  "devDependencies": {
+    "@types/react": "^16.9.34",
+    "@typescript-eslint/eslint-plugin": "^2.29.0",
+    "@typescript-eslint/parser": "^2.29.0",
+    "eslint": "^6.8.0",
+    "prettier": "^2.0.5",
+    "typescript": "^3.8.3"
+  },
+  "dependencies": {
+    "@types/node-fetch": "^2.5.7",
+    "@urql/core": "^1.11.7",
+    "graphql": "^15.0.0",
+    "graphql-tag": "^2.10.3",
+    "isomorphic-unfetch": "^3.0.0",
+    "node-fetch": "^2.6.0",
+    "plumbery-core": "*",
+    "wonka": "^4.0.9"
+  }
+}
