+++ packages/connector-thegraph-voting/src/index.ts
@@ -1,0 +1,5 @@
+export { default as VotingConnectorTheGraph } from './connector'
+
+export { default as Voting } from './entities/Voting'
+export { default as VotingCast } from './entities/VotingCast'
+export { default as VotingVote } from './entities/VotingVote'
