+++ packages/connector-thegraph-voting/src/parsers/index.ts
@@ -1,0 +1,2 @@
+export { parseVotes } from './votes'
+export { parseCasts } from './casts'
