+++ packages/connector-thegraph-voting/src/parsers/votes.ts
@@ -1,0 +1,19 @@
+import { Vote as VoteDataGql } from '../queries/types'
+import { VoteData } from '../entities/VotingVote'
+import { QueryResult } from 'plumbery-connector-thegraph'
+
+export function parseVotes(
+  result: QueryResult
+): VoteData[] {
+  const votes = result.data.votes
+
+  if (!votes) {
+    throw new Error('Unable to parse votes.')
+  }
+
+  // Note, this may seem redundant, but it makes sure
+  // types are enforced.
+  return votes.map((vote: VoteDataGql): VoteData => {
+    return vote
+  })
+}
