--- packages/connector-thegraph/src/voting/parsers/casts.ts
@@ -1,22 +1,0 @@
-import { Cast as CastDataGql } from '../queries/types'
-import { CastData } from '../entities/VotingCast'
-import { QueryResult } from '../../types'
-
-export function parseCasts(
-  result: QueryResult
-): CastData[] {
-  const casts = result.data.casts
-
-  if (!casts) {
-    throw new Error('Unable to parse casts.')
-  }
-
-  return casts.map((cast: CastDataGql): CastData => {
-    return {
-      id: cast.id,
-      voteId: cast.voteId,
-      voter: cast.voter,
-      supports: cast.supports
-    }
-  })
-}
