--- packages/connector-thegraph/src/voting/parsers/index.ts
@@ -1,2 +1,0 @@
-export { parseVotes } from './votes'
-export { parseCasts } from './casts'
